Refactor: Improve file structure and configuration portability
I've moved all Python source files into a `src` directory with subdirectories for bot, AI, external application handling, and utilities. This provides a cleaner and more organized project layout.

Key changes:
- I created a `src/` directory as the main container for source code.
- I organized modules into `src/bot/`, `src/ai/`, `src/external/`, and `src/utils/`.
- I updated all import statements across the project to reflect the new structure.
- I added `__init__.py` files to `src/` and its subdirectories to ensure they are treated as Python packages.
- I modified `config.yaml` to use relative paths for document uploads (e.g., `sample_resume.pdf`) and the output directory (`output/`) instead of absolute, system-specific paths.
- I created placeholder files (`sample_resume.pdf`, `sample_text_resume.txt`, `sample_docx_resume.docx`) and an `output/.placeholder` to support the new relative paths in `config.yaml` and improve testability/portability.

These changes make the project easier to navigate, maintain, and run in different environments.
