+++ output/.placeholder
@@ -1,0 +1,1 @@
+# This placeholder creates the output directory.
