+++ sample_docx_resume.docx
@@ -1,0 +1,1 @@
+This is a placeholder DOCX resume file.
