+++ sample_resume.pdf
@@ -1,0 +1,1 @@
+This is a placeholder PDF resume file.
