+++ sample_text_resume.txt
@@ -1,0 +1,1 @@
+This is a placeholder text resume file.
