+++ src/ai/__init__.py
@@ -1,0 +1,1 @@
+# This file makes ai a package.
