+++ src/bot/__init__.py
@@ -1,0 +1,1 @@
+# This file makes bot a package.
