+++ src/external/.placeholder
@@ -1,0 +1,1 @@
+# This is a placeholder file to create the external directory.
