+++ src/external/__init__.py
@@ -1,0 +1,1 @@
+# This file makes external a package.
